Remove auto save/backup settings which break editing of Dropbox files
This config caused many headaches when working with org files stored in
Dropbox.

What appeared to be happening was that Emacs would be editing a
temporary file, Dropbox would look like it was syncing the file changes
but in actual fact wasn't syncing anything. For the first few seconds
you would be editing the actual file, and then emacs would switch to
editing a temporary file (presumably a backup, or autosave file).